+++ amstrax/plugins/events/__init__.py
@@ -1,2 +1,19 @@
-from . import event_processing
-from .event_processing import *
+from . import events
+from .events import *
+
+from . import event_basics
+from .event_basics import *
+
+from . import event_positions
+from .event_positions import *
+
+from . import corrected_areas
+from .corrected_areas import *
+
+from . import energy_estimates
+from .energy_estimates import *
+
+from . import event_info
+from .event_info import *
+
+
