+++ amstrax/plugins/peaks/peak_classification.py
@@ -1,6 +1,7 @@
 import numba
 import numpy as np
 import strax
+export, __all__ = strax.exporter()
 
 @export
 @strax.takes_config(
